Move full post header into reader-full-post component
