+++ assets/stylesheets/_components.scss
@@ -95,7 +95,7 @@
 @import 'components/purchase-detail/style';
 @import 'components/push-notification/style';
 @import 'components/rating/style';
-@import 'components/reader-full-post-header/style';
+@import 'components/reader-full-post/style';
 @import 'components/related-posts/style';
 @import 'components/ribbon/style';
 @import 'components/search/style';
